Fixed a few bugs and did some cleanup for joining.
- Y changes not been detected
- Match stop when a team has over 50% of
the total rounds won.
- Player should not be  able to bend before match
starts!
