+++ src/com/projectkorra/probending/commands/PBSubCommand.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.bukkit.command.CommandSender;
 
-import com.projectkorra.probending.managers.FieldCreationManager;
+import com.projectkorra.probending.managers.PBFieldCreationManager;
 import com.projectkorra.probending.managers.ProbendingHandler;
 import com.projectkorra.projectkorra.command.PKCommand;
 
@@ -54,5 +54,5 @@ public interface PBSubCommand {
 	 * @param sender The CommandSender who issued the command
 	 * @param args the command's arguments
 	 */
-	public void execute(CommandSender sender, List<String> args, ProbendingHandler pHandler, FieldCreationManager cManager);
+	public void execute(CommandSender sender, List<String> args);
 }
