+++ src/com/projectkorra/probending/commands/QueueJoinCommand.java
@@ -8,8 +8,6 @@ import org.bukkit.entity.Player;
 import com.projectkorra.probending.PBMessenger;
 import com.projectkorra.probending.PBMessenger.PBMessage;
 import com.projectkorra.probending.enums.GamePlayerMode;
-import com.projectkorra.probending.managers.FieldCreationManager;
-import com.projectkorra.probending.managers.ProbendingHandler;
 
 import net.md_5.bungee.api.ChatColor;
 
@@ -20,7 +18,7 @@ public class QueueJoinCommand extends PBCommand {
 	}
 
 	@Override
-	public void execute(CommandSender sender, List<String> args, ProbendingHandler pHandler, FieldCreationManager cManager) {
+	public void execute(CommandSender sender, List<String> args) {
 		if (!sender.hasPermission("probending.command.join")) {
 			sender.sendMessage(ChatColor.RED + "Insufficient Permissions");
 			return;
@@ -29,14 +27,14 @@ public class QueueJoinCommand extends PBCommand {
 			Player player = (Player) sender;
 			if (args.size() > 0) {
 				if (args.get(0).equalsIgnoreCase("1")) {
-					pHandler.quePlayer(player, GamePlayerMode.SINGLE);
+					Commands.getqManager().quePlayer(player, GamePlayerMode.SINGLE);
 				} else if (args.get(0).equalsIgnoreCase("3")) {
-					pHandler.quePlayer(player, GamePlayerMode.TRIPLE);
+					Commands.getqManager().quePlayer(player, GamePlayerMode.TRIPLE);
 				} else {
 					PBMessenger.sendMessage(player, PBMessage.ERROR);
 				}
 			} else {
-				pHandler.quePlayer(player, GamePlayerMode.ANY);
+				Commands.getqManager().quePlayer(player, GamePlayerMode.ANY);
 			}
 		} else {
 			sender.sendMessage(ChatColor.RED + "Only players can use this command!");
