+++ src/com/projectkorra/probending/commands/QueueLeaveCommand.java
@@ -7,7 +7,7 @@ import org.bukkit.entity.Player;
 
 import com.projectkorra.probending.PBMessenger;
 import com.projectkorra.probending.PBMessenger.PBMessage;
-import com.projectkorra.probending.managers.FieldCreationManager;
+import com.projectkorra.probending.managers.PBFieldCreationManager;
 import com.projectkorra.probending.managers.ProbendingHandler;
 
 import net.md_5.bungee.api.ChatColor;
@@ -19,7 +19,7 @@ public class QueueLeaveCommand extends PBCommand{
 	}
 
 	@Override
-	public void execute(CommandSender sender, List<String> args, ProbendingHandler pHandler, FieldCreationManager cManager) {
+	public void execute(CommandSender sender, List<String> args) {
 		if (!sender.hasPermission("probending.command.leave")) {
 			sender.sendMessage(ChatColor.RED + "Insufficient Permissions");
 			return;
@@ -29,7 +29,7 @@ public class QueueLeaveCommand extends PBCommand{
 			if (args.size() > 0) {
 				PBMessenger.sendMessage(player, PBMessage.ERROR);
 			} else {
-				pHandler.removePlayerFromQueue(player);
+				Commands.getqManager().removePlayerFromQueue(player);
 			}
 		} else {
 			sender.sendMessage(ChatColor.RED + "Only players can use this command!");
